Refactor interaction database management and schema organization
- Modularized the database interaction logic by separating schema, row mapping, and spool functionalities into distinct modules for improved maintainability.
- Updated the interaction spool initialization to utilize a centralized schema definition, enhancing clarity and reducing redundancy.
- Implemented row mapping functions for sessions, turns, and events to streamline data retrieval and manipulation.
- Added comprehensive tests to validate interaction recording and checkpoint assignment, ensuring robust functionality across the interaction lifecycle.
